+++ packages/codemod/src/50-slider-input/__testfixtures__/transform.output.tsx
@@ -1,0 +1,19 @@
+import React, {Fragment} from 'react';
+
+import { SliderInput } from '@alfalab/core-components-slider-input';
+
+export const Component = () => (
+    <Fragment>
+        <SliderInput size={48} />
+        <SliderInput size={48} />
+
+        <SliderInput size={56} />
+        <SliderInput size={56} />
+
+        <SliderInput size={64} />
+        <SliderInput size={64} />
+
+        <SliderInput size={72} />
+        <SliderInput size={72} />
+    </Fragment>
+);