+++ packages/file-upload-item/src/components/content/content.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import cn from 'classnames';
 
-import { Text } from '@alfalab/core-components-typography';
+import { TypographyText } from '@alfalab/core-components-typography';
 
 import { FileUploadItemContext } from '../../context/file-upload-item-context';
 
@@ -32,16 +32,16 @@ export const Content = () => {
             })}
         >
             {title && (
-                <Text
+                <TypographyText
                     className={cn(styles.title, {
                         [styles.truncate]: truncate,
                         [styles.restore]: showRestore,
                     })}
-                    view='component'
+                    view='component-primary'
                     color='primary'
                 >
                     {title}
-                </Text>
+                </TypographyText>
             )}
             <ContentSubtitle />
         </div>