+++ packages/pure-cell/src/components/text/component.tsx
@@ -2,7 +2,7 @@ import React, { type ReactNode, useContext } from 'react';
 import cn from 'classnames';
 
 import { getDataTestId } from '@alfalab/core-components-shared';
-import { type Color, Text as TypographyText } from '@alfalab/core-components-typography';
+import { type Color, TypographyText } from '@alfalab/core-components-typography';
 
 import { PureCellContext } from '../../context';
 
@@ -26,8 +26,8 @@ type Props = {
         | 'primary-large'
         | 'primary-medium'
         | 'primary-small'
-        | 'component'
         | 'component-primary'
+        | 'component-secondary'
         | 'secondary-large';
 
     /**