+++ packages/select/src/components/options-list/Component.tsx
@@ -4,7 +4,7 @@ import cn from 'classnames';
 
 import { Scrollbar } from '@alfalab/core-components-scrollbar';
 
-import { DEFAULT_VISIBLE_OPTIONS, SIZE_TO_CLASSNAME_MAP } from '../../consts';
+import { DEFAULT_VISIBLE_OPTIONS } from '../../consts';
 import { useNativeScrollbar } from '../../hooks/use-native-scrollbar';
 import { type GroupShape, type OptionShape, type OptionsListProps } from '../../typings';
 import { isGroup, useVisibleOptions } from '../../utils';
@@ -204,7 +204,7 @@ export const OptionsList = forwardRef<HTMLDivElement, OptionsListProps>(
         return (
             <div
                 {...(nativeScrollbar && { 'data-test-id': dataTestId })}
-                className={cn(styles.optionsList, styles[SIZE_TO_CLASSNAME_MAP[size]], className)}
+                className={cn(styles.optionsList, styles[`size-${size}`], className)}
             >
                 {header && (
                     <div