+++ packages/toast-plate/src/components/base-toast-plate/utils/isUnsafeBadge.ts
@@ -1,9 +1,0 @@
-import { type StatusBadgeProps } from '@alfalab/core-components-status-badge';
-
-import { type unsafe_BadgeProps } from '../types/unsafeBadgeProps';
-
-const unsafeBadgeValues: string[] = ['negative', 'positive', 'attention'];
-
-type IsUnsafeBadge = unsafe_BadgeProps | StatusBadgeProps['view'];
-export const isUnsafeBadge = (badge: IsUnsafeBadge): badge is unsafe_BadgeProps =>
-    unsafeBadgeValues.includes(badge);